ReactJS: Move OutputStreamToSLF4J into the utils sub-package.
This class is generally useful and could be a part of Micronaut (or really, SLF4J itself).
