+++ views-react/src/main/java/io/micronaut/views/react/CompiledJS.java
@@ -20,6 +20,7 @@ import io.micronaut.core.annotation.Internal;
 import io.micronaut.scheduling.io.watch.event.FileChangedEvent;
 import io.micronaut.scheduling.io.watch.event.WatchEventType;
 import io.micronaut.views.react.util.BeanPool;
+import io.micronaut.views.react.util.OutputStreamToSLF4J;
 import jakarta.annotation.PreDestroy;
 import jakarta.inject.Inject;
 import jakarta.inject.Singleton;
