+++ views-react/src/main/java/io/micronaut/views/react/util/OutputStreamToSLF4J.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.views.react;
+package io.micronaut.views.react.util;
 
 import io.micronaut.core.annotation.Internal;
 import io.micronaut.core.annotation.NonNull;
@@ -30,7 +30,7 @@ import java.nio.charset.Charset;
  * An output stream that looks for line separators and then writes out the lines of text to the given logger.
  */
 @Internal
-final class OutputStreamToSLF4J extends OutputStream {
+public final class OutputStreamToSLF4J extends OutputStream {
     private final Charset charset;
 
     private ByteBuffer buffer = ByteBuffer.allocate(512);
